# backend/docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CheckpointDto:
    properties:
      ack:
        description: Checkpoint ID
        type: string
      type:
        description: Sync type (e.g. "assets_v1")
        type: string
    type: object
...
  dto.DeleteCheckpointRequest:
    properties:
      types:
        description: Sync types to clear
        items:
          type: string
        minItems: 1
        type: array
    required:
    - types
    type: object
...
  dto.GetCheckpointResponse:
    properties:
      checkpoints:
        items:
          $ref: '#/definitions/dto.CheckpointDto'
        type: array
    type: object
...
  dto.SetCheckpointRequest:
    properties:
      checkpoints:
        items:
          $ref: '#/definitions/dto.CheckpointDto'
        minItems: 1
        type: array
    required:
    - checkpoints
    type: object
  dto.SyncRequestType:
    enum:
    - assets_v1
    type: string
    x-enum-varnames:
    - SyncRequestTypeAssetsV1
  dto.SyncStreamRequest:
    properties:
      reset:
        type: boolean
      types:
        items:
          $ref: '#/definitions/dto.SyncRequestType'
        minItems: 1
        type: array
      updated_after:
        description: RFC3339 timestamp
        type: string
    required:
    - types
    type: object
...
  response.ApiResponse:
    properties:
      code:
        description: 0 success, non-zero failure
        type: integer
      data:
        description: Optional payload
      message:
        description: Human-readable message
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: PrismBox backend HTTP API reference
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: PrismBox Backend API
  version: "1.0"
paths:
...
  /sync/assets/stream:
    post:
      consumes:
      - application/json
      description: 'Full or incremental sync. On incremental sync the server may emit
        `asset_delete_v1` after asset rows so clients can set deletedAt locally. Example
        delete event: {"type":"asset_delete_v1","ids":["uuid1"],"data":{}}.'
      parameters:
      - description: Sync request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SyncStreamRequest'
      produces:
      - application/jsonlines+json
      responses:
        "200":
          description: 'JSON Lines stream: asset_v1, asset_delete_v1, sync_complete_v1'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Stream asset sync
      tags:
      - Sync
  /sync/checkpoint:
    delete:
      consumes:
      - application/json
      description: Deletes checkpoints for the listed sync types on this device
      parameters:
      - description: Delete payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.DeleteCheckpointRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No content
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Delete checkpoints
      tags:
      - Sync
    get:
      description: Checkpoints for the current user and device
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.GetCheckpointResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Get checkpoints
      tags:
      - Sync
    post:
      consumes:
      - application/json
      description: Upserts checkpoint ack values for the device
      parameters:
      - description: Checkpoint payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SetCheckpointRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No content
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Set checkpoints
      tags:
      - Sync
...
securityDefinitions:
  BearerAuth:
    description: 'Use `Authorization: Bearer {token}` or header `x-prismbox-user-token:
      {token}`'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"